Add files via upload
The files in this archive have names in the format S#P#, which stands for Sketch # Point #. These names reference the sketch number in the book, Remote Sensor Monitoring by Radio with Arduino: Detecting Intruders, Fires, Flammable and Toxic Gases, and Other Hazards at a Distance. For example, a file called S3P1 would be Sketch 3.1 in the book.

Copy or move the RemoteSensors folder to your Documents/Arduino folder and it should show up in Sketchbook under the Arduino File menu. You may need to shut down and restart the IDE if it was running when you added the RemoteSensors folder to the Documents/Arduino folder.